requirements.txt
pyyaml
websocket-client
requests